增加了成果列表新页面 增加了成果列表对应数据的数据库  增加了成果申报的页面跳转

+++ rate/rateserver/rate-mapper/src/main/java/org/sys/rate/mapper/PatentMapper.java
@@ -26,7 +26,7 @@ public interface PatentMapper
      * @return
      */
     @Select("select * from i_patent where ID = #{ID}")
-    Patent getById(Integer ID);
+    Patent getById(Long ID);
 
     /**
      * 查询专利成果
